feat: event source stream file rename
